Cleaner split of tests and validation
